*short: Speedmeter time window adjustable*
Adjust the time period for the speedmeter from 10 to *** seconds

*codelevel: Addon (2)*
-----------------------------------------------
*short: browser.submitForm fix for image inputfields*
imageinputfields without name are not sent any more. Before the got send with null.x=54 and null.y=43

*codelevel: Libraries(6)*

*May affect*
All caller of browser.submit that  use this auto image feature.

*testing*
I guess that this feature has not been used yet. So there shouldn't be sideffects


git-svn-id: svn://svn.jdownloader.org/jdownloader@4753 ebf7c1c2-ba36-0410-9fe8-c592906822b4